refactor(trusty-search): split oversized reindex_engine under the 500-line cap
Decompose the 1,877-line commands/reindex_engine.rs into a focused module
directory with a thin re-export facade, per the 500-line file cap. Pure
mechanical move — no behavior change, no signature change. All 13 existing
reindex_engine unit tests pass unchanged.

New module layout (crates/trusty-search/src/commands/reindex_engine/):
- mod.rs           facade re-exporting the prior public surface (60 lines)
- file_ops.rs      index_single_file, add_path (87)
- options.rs       ReindexOptions, ReindexOutcome (97)
- registration.rs  RegisterFilters, register_index_with_daemon{,_filtered},
                   fetch_chunk_count (159)
- driver.rs        run_reindex{,_opts,_force_opts}, run_reindex_with (461)
- events.rs        SSE event dispatch + LoopState (359)
- event_handlers.rs handle_batch / handle_complete (202)
- ticker.rs        wall-clock stats-line ticker (97)
- phase_map.rs     phase discriminant encode/decode (54)
- progress_state.rs SharedProgress atomic bundle (65)
- verify.rs        post-force health check (89)
- tests.rs         the 13 unit tests, moved verbatim (455)

Every resulting .rs file is under the cap. Removes the grandfathered
reindex_engine.rs entry from .line-cap-allowlist.tsv (ratchet-down 168→167).

Gates: cargo check (workspace) clean; clippy -p trusty-search --all-targets
-D warnings clean; cargo fmt --check clean; cargo test -p trusty-search pass;
scripts/check_line_cap.sh exit 0.

Refs #990 (trusty-search server.rs already split in #828; trusty-memory web.rs
and trusty-mpm tm.rs are out of scope for this trusty-search-only PR).

Closes #571
