+++ .line-cap-allowlist.tsv
@@ -135,7 +135,6 @@ crates/trusty-review/src/pipeline/runner_tests.rs	738
 crates/trusty-review/src/pipeline/verify_tests.rs	556
 crates/trusty-search/src/commands/doctor_checks.rs	612
 crates/trusty-search/src/commands/integrate.rs	688
-crates/trusty-search/src/commands/reindex_engine.rs	1877
 crates/trusty-search/src/commands/reindex_ui.rs	1468
 crates/trusty-search/src/commands/start.rs	1889
 crates/trusty-search/src/core/chunker/mod.rs	1883
